# Update go to v1.24.0