<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->
**What this PR does**:

Continues https://github.com/grafana/tempo/pull/4685

1. Update go to v1.24.0
2. Use new version on linter image that supports go1.24
3. Minor fixes for go update

**Checklist**

- [ ] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`